rdrf/runtime-requirements.txt
Django==1.8.8
pymongo==2.8.1
pyyaml
django-extensions>=0.7.1
django-picklefield==0.1.9
django-templatetag-sugar==1.0
pyparsing==1.5.7
wsgiref==0.1.2
python-memcached==1.54
django-messages-ui==0.2.7
django-nose
django-templatetag-handlebars==1.2.0
django-ajax-selects
nose
sphinxcontrib-httpdomain==1.4.0
setuptools_scm==1.10.1
hgtools==6.5.1
django-positions 
pycountry==1.12
django-countries
uwsgi==2.0.8
uwsgitop
pyinotify==0.9.4
six==1.9.0
python-gettext
django-registration-redux==1.3
psycopg2==2.6.1
django-secure==1.0.1
djangorestframework==3.3.3
https://github.com/muccg/ccg-django-utils/archive/0.3.1.tar.gz
openpyxl
SQLAlchemy
django-useraudit
django-iprestrict
django_mailgun
